+++ core/src/dird/dird.cc
@@ -27,6 +27,7 @@
  */
 
 #include "include/bareos.h"
+#include "include/exit_codes.h"
 #include "cats/sql.h"
 #include "cats/sql_pooling.h"
 #include "dird.h"
@@ -226,7 +227,7 @@ int main(int argc, char* argv[])
 
   AddDeprecatedExportOptionsHelp(dir_app);
 
-  CLI11_PARSE(dir_app, argc, argv);
+  ParseBareosApp(dir_app, argc, argv);
 
   if (!no_signals) { InitSignals(TerminateDird); }
 
@@ -251,18 +252,18 @@ int main(int argc, char* argv[])
             "but program was not started with required root privileges.\n"));
   }
 
-  my_config = InitDirConfig(configfile, M_ERROR_TERM);
+  my_config = InitDirConfig(configfile, M_CONFIG_ERROR);
   if (export_config_schema) {
     PoolMem buffer;
 
     PrintConfigSchemaJson(buffer);
     printf("%s\n", buffer.c_str());
 
-    TerminateDird(0);
-    return 0;
+    TerminateDird(BEXIT_SUCCESS);
+    return BEXIT_SUCCESS;
   }
 
-  my_config->ParseConfig();
+  my_config->ParseConfigOrExit();
 
   if (export_config) {
     int rc = 0;
@@ -272,7 +273,7 @@ int main(int argc, char* argv[])
       rc = 1;
     }
     TerminateDird(rc);
-    return 0;
+    return BEXIT_SUCCESS;
   }
 
   if (!CheckResources()) {
@@ -280,8 +281,8 @@ int main(int argc, char* argv[])
          _("Please correct the configuration in %s\n"),
          my_config->get_base_config_path().c_str());
 
-    TerminateDird(0);
-    return 0;
+    TerminateDird(BEXIT_SUCCESS);
+    return BEXIT_SUCCESS;
   }
 
   if (my_config->HasWarnings()) {
@@ -301,8 +302,8 @@ int main(int argc, char* argv[])
     Jmsg((JobControlRecord*)nullptr, M_ERROR_TERM, 0,
          _("Cryptography library initialization failed.\n"));
 
-    TerminateDird(0);
-    return 0;
+    TerminateDird(BEXIT_SUCCESS);
+    return BEXIT_SUCCESS;
   }
 
   if (!test_config) {
@@ -323,8 +324,8 @@ int main(int argc, char* argv[])
          _("Please correct the configuration in %s\n"),
          my_config->get_base_config_path().c_str());
 
-    TerminateDird(0);
-    return 0;
+    TerminateDird(BEXIT_SUCCESS);
+    return BEXIT_SUCCESS;
   }
 
   if (test_config) { TerminateDird(0); }
@@ -334,8 +335,8 @@ int main(int argc, char* argv[])
          _("Please correct the configuration in %s\n"),
          my_config->get_base_config_path().c_str());
 
-    TerminateDird(0);
-    return 0;
+    TerminateDird(BEXIT_SUCCESS);
+    return BEXIT_SUCCESS;
   }
 
   MyNameIs(0, nullptr, me->resource_name_); /* set user defined name */
@@ -370,8 +371,8 @@ int main(int argc, char* argv[])
 
   Scheduler::GetMainScheduler().Run();
 
-  TerminateDird(0);
-  return 0;
+  TerminateDird(BEXIT_SUCCESS);
+  return BEXIT_SUCCESS;
 }
 
 /**
@@ -389,7 +390,7 @@ static
 
   if (is_reloading) {  /* avoid recursive termination problems */
     Bmicrosleep(2, 0); /* yield */
-    exit(1);
+    exit(BEXIT_FAILURE);
   }
 
   is_reloading = true;