+++ core/src/tools/drivetype.cc
@@ -27,10 +27,11 @@
 
 #include <unistd.h>
 #include "include/bareos.h"
+#include "include/exit_codes.h"
 #include "findlib/find.h"
 #include "findlib/drivetype.h"
 
-static void usage()
+static void usage(int exit_code)
 {
   fprintf(stderr,
           _("\n"
@@ -45,13 +46,13 @@ static void usage()
             "       -?     print this message.\n"
             "\n"));
 
-  exit(1);
+  exit(exit_code);
 }
 
 int DisplayDrive(char* drive, bool display_local, int verbose)
 {
   char dt[100];
-  int status = 0;
+  int status = BEXIT_SUCCESS;
 
   if (Drivetype(drive, dt, sizeof(dt))) {
     if (display_local) { /* in local mode, display only harddrive */
@@ -63,7 +64,7 @@ int DisplayDrive(char* drive, bool display_local, int verbose)
     }
   } else if (!display_local) { /* local mode is used by FileSet scripts */
     fprintf(stderr, _("%s: unknown\n"), drive);
-    status = 1;
+    status = BEXIT_FAILURE;
   }
   return status;
 }
@@ -71,7 +72,6 @@ int DisplayDrive(char* drive, bool display_local, int verbose)
 int main(int argc, char* const* argv)
 {
   int verbose = 0;
-  int status = 0;
   int ch, i;
   bool display_local = false;
   bool display_all = false;
@@ -95,8 +95,10 @@ int main(int argc, char* const* argv)
         display_all = true;
         break;
       case '?':
+        usage(BEXIT_SUCCESS);
+        break;
       default:
-        usage();
+        usage(BEXIT_FAILURE);
     }
   }
   argc -= optind;
@@ -104,19 +106,20 @@ int main(int argc, char* const* argv)
 
   OSDependentInit();
 
-  if (argc < 1 && display_all) {
+  if (display_all) {
     /* Try all letters */
     for (drive = 'A'; drive <= 'Z'; drive++) {
       Bsnprintf(buf, sizeof(buf), "%c:/", drive);
       DisplayDrive(buf, display_local, verbose);
     }
-    exit(status);
+    exit(BEXIT_SUCCESS);
   }
 
-  if (argc < 1) { usage(); }
+  if (argc < 1) { usage(BEXIT_FAILURE); }
 
+  int exit_status = BEXIT_SUCCESS;
   for (i = 0; i < argc; --argc, ++argv) {
-    status += DisplayDrive(*argv, display_local, verbose);
+    exit_status = DisplayDrive(*argv, display_local, verbose);
   }
-  exit(status);
+  exit(exit_status);
 }